refactor(SFOX Bank Link): move plaid information into selectors file
